test: improve unit testing
